+++ pkg/dds/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 4.2.4
 - Added missing type to `Event` in `postEvent`.
+- [DAP] Instaces with both fields and getters of the same name will no longer show duplicates in `variables` responses.
 
 # 4.2.3
 - Added missing await of `WebSocketChannel.ready` in `startDartDevelopmentService`.